FEATURE: Create notification when a comment is created. (#55)
When a comment is created, we want to alert that creator of the post
where the comment is made. We may want to consider consolidatig
notifications in the future but this will do for the initial version of
the plugin.

To link to the comment, we rely on a hash fragment in the URL. Ideally,
we want to disable the post's highlight and highlight the comment
instead. However, this requires alot of hacking and our focus now is to
get an MVP out. If we really want to sugar coat this in the future, it
is still possible.